+++ migration/003-create_consumer_limits_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS `consumer_limits`(
     `consumer_limit_id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `consumer_id` BIGINT UNSIGNED NOT NULL,
     `tenure` ENUM('1', '2', '3', '6') NOT NULL,
-    `limit_amount` DOUBLE NOT NULL,
+    `limit_amount` DECIMAL(19, 3) NOT NULL,
     `created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     `deleted_at` TIMESTAMP NULL,
